requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
idna==2.10
msedge-selenium-tools==3.141.3
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
